pyproject.toml
[project]
name = "advanced-rag-showcase"
version = "0.1.0"
description = "Advanced Retrieval-Augmented Generation (RAG) experiments with hybrid retrieval, cross-encoder reranking, and multiple LLM providers via LangChain."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "langchain>=0.3.27",
    "openai>=1.68.2",
    "anthropic>=0.6.1",
    "google-generativeai>=0.2.0",
    "groq>=0.11.0",
    "datasets>=2.20.0",
    "pandas>=2.2.2",
    "tqdm>=4.66.4",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "rank_bm25>=0.2.1",
    "transformers>=4.35.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.102.0",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.11.7",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "pypdf>=6.0.0",
    "unstructured>=0.18.14",
    "python-docx>=1.2.0",
    "Markdown>=3.9",
    "langchain-huggingface>=0.3.1",
    "ipywidgets>=8.1.7",
    "langchain-cohere>=0.4.5",
    "chromadb>=1.0.20",
    "langchain-weaviate>=0.0.5",
    "langchain-pinecone>=0.2.12",
    "langchain-milvus>=0.2.1",
    "pymilvus>=2.6.1",
    "gradio>=5.45.0",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.5.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"
